+++ Catty/Sprite.h
@@ -97,8 +97,8 @@
 - (void)addSound:(AVAudioPlayer*)sound;
 - (void)comeToFront;
 - (void)changeSizeByN:(float)sizePercentageRate;
-- (void)changeXBy:(float)x;
-- (void)changeYBy:(float)y;
+- (void)changeXBy:(int)x;
+- (void)changeYBy:(int)y;
 - (void)stopAllSounds;
 - (void)setSizeToPercentage:(float)sizeInPercentage;
 //- (void)addLoopBricks:(NSArray*)bricks;
